Add Fixture protocol for sample data, empty out DefaultDatabase reads
DefaultDatabase.read now returns an empty chunk instead of serving
RecordDecodable.samples, decoupling sample data from the record-reading
path. Types that expose sample data now conform to a new Fixture
protocol (Sources/Scout/UI/Database/Fixture.swift), which provides
[Type].samples and bare .samples through a single generic bridge,
independent of RecordDecodable. Types with more than one sample-related
helper (Crash, Hang, Event, Device, Install, Launch, Session, Version,
ParamProvider.Item) now group their helpers in a dedicated
Type+Fixture.swift file next to the type's main file.

Updated CLAUDE.md's Sample data section to document the new convention.
