+++ Sources/Scout/UI/Home/Section/Release/HomeReleaseSection.swift
@@ -46,7 +46,7 @@ struct HomeReleaseSection: View {
 
 #Preview {
     let provider = ReleaseHealthProvider()
-    provider.result = .success(ReleaseHealth.samples)
+    provider.result = .success(.samples)
 
     return NavigationStack {
         List {
