+++ Sources/Scout/UI/Network/View/NetworkEndpointRow.swift
@@ -61,7 +61,7 @@ struct NetworkEndpointRow: View {
 }
 
 #Preview("NetworkEndpointRow") {
-    List(NetworkEndpoint.samples) { endpoint in
+    List([NetworkEndpoint].samples) { endpoint in
         NetworkEndpointRow(endpoint: endpoint)
     }
     .listStyle(.plain)
