+++ Sources/Scout/UI/Settings/BackendDetailView.swift
@@ -167,11 +167,11 @@ struct PingSparkline: View {
 
 @available(iOS 17.0, macOS 14.0, *)
 #Preview("Server") {
-    @Previewable @State var activeID = BackendHealth.samples[0].id
+    @Previewable @State var activeID = [BackendHealth].samples[0].id
     NavigationStack {
         BackendDetailView(
-            provider: BackendHealthProvider(healths: BackendHealth.samples),
-            id: BackendHealth.samples[0].id,
+            provider: BackendHealthProvider(healths: .samples),
+            id: [BackendHealth].samples[0].id,
             activeID: $activeID
         )
     }
@@ -179,11 +179,11 @@ struct PingSparkline: View {
 
 @available(iOS 17.0, macOS 14.0, *)
 #Preview("CloudKit") {
-    @Previewable @State var activeID = BackendHealth.samples[0].id
+    @Previewable @State var activeID = [BackendHealth].samples[0].id
     NavigationStack {
         BackendDetailView(
-            provider: BackendHealthProvider(healths: BackendHealth.samples),
-            id: BackendHealth.samples[1].id,
+            provider: BackendHealthProvider(healths: .samples),
+            id: [BackendHealth].samples[1].id,
             activeID: $activeID
         )
     }
@@ -191,11 +191,11 @@ struct PingSparkline: View {
 
 @available(iOS 17.0, macOS 14.0, *)
 #Preview("Unreachable") {
-    @Previewable @State var activeID = BackendHealth.samples[0].id
+    @Previewable @State var activeID = [BackendHealth].samples[0].id
     NavigationStack {
         BackendDetailView(
-            provider: BackendHealthProvider(healths: BackendHealth.samples),
-            id: BackendHealth.samples[3].id,
+            provider: BackendHealthProvider(healths: .samples),
+            id: [BackendHealth].samples[3].id,
             activeID: $activeID
         )
     }
