+++ Sources/Scout/UI/Timeline/Rail/Launch.swift
@@ -25,14 +25,6 @@ extension Launch: RecordDecodable {
         "install_id",
     ]
 
-    static var samples: [Launch] {
-        let installID = UUID()
-        return [
-            .sample(minutesAgo: 0, installID: installID, ongoing: true),
-            .sample(minutesAgo: 90, installID: installID),
-        ]
-    }
-
     init(record: Record) throws {
         startDate = record["start_date"]
         endDate = record["end_date"]
@@ -50,18 +42,5 @@ extension Launch: RecordEncodable {
         record["launch_id"] = launchID?.uuidString
         record["install_id"] = installID?.uuidString
         return record
-    }
-}
-
-extension Launch {
-    static func sample(minutesAgo: Double = 0, launchID: UUID = UUID(), installID: UUID = UUID(), ongoing: Bool = false) -> Launch {
-        let startDate = Date(timeIntervalSinceNow: -minutesAgo * 60)
-        return Launch(
-            startDate: startDate,
-            endDate: ongoing ? nil : startDate.addingTimeInterval(180),
-            id: launchID.uuidString,
-            launchID: launchID,
-            installID: installID
-        )
     }
 }
