+++ Sources/Scout/UI/Timeline/Rail/Session.swift
@@ -27,15 +27,6 @@ extension Session: RecordDecodable {
         "install_id",
     ]
 
-    static var samples: [Session] {
-        let launchID = UUID()
-        let installID = UUID()
-        return [
-            .sample(minutesAgo: 0, launchID: launchID, installID: installID, ongoing: true),
-            .sample(minutesAgo: 45, launchID: launchID, installID: installID),
-        ]
-    }
-
     init(record: Record) throws {
         startDate = record["start_date"]
         endDate = record["end_date"]
@@ -55,19 +46,5 @@ extension Session: RecordEncodable {
         record["launch_id"] = launchID?.uuidString
         record["install_id"] = installID?.uuidString
         return record
-    }
-}
-
-extension Session {
-    static func sample(minutesAgo: Double = 0, sessionID: UUID = UUID(), launchID: UUID = UUID(), installID: UUID = UUID(), ongoing: Bool = false) -> Session {
-        let startDate = Date(timeIntervalSinceNow: -minutesAgo * 60)
-        return Session(
-            startDate: startDate,
-            endDate: ongoing ? nil : startDate.addingTimeInterval(120),
-            id: sessionID.uuidString,
-            sessionID: sessionID,
-            launchID: launchID,
-            installID: installID
-        )
     }
 }
