+++ Sources/Scout/UI/Timeline/Rail/Version.swift
@@ -25,13 +25,6 @@ extension Version: RecordDecodable {
         "date",
     ]
 
-    static var samples: [Version] {
-        [
-            .sample(appVersion: "2.4.1", buildNumber: "214", minutesAgo: 0),
-            .sample(appVersion: "2.4.0", buildNumber: "210", minutesAgo: 4320),
-        ]
-    }
-
     init(record: Record) throws {
         appVersion = record["app_version"]
         buildNumber = record["build_number"]
@@ -49,17 +42,5 @@ extension Version: RecordEncodable {
         record["launch_id"] = launchID?.uuidString
         record["date"] = date
         return record
-    }
-}
-
-extension Version {
-    static func sample(appVersion: String = "2.4.1", buildNumber: String = "214", minutesAgo: Double = 0, launchID: UUID = UUID()) -> Version {
-        Version(
-            appVersion: appVersion,
-            buildNumber: buildNumber,
-            launchID: launchID,
-            date: Date(timeIntervalSinceNow: -minutesAgo * 60),
-            id: UUID().uuidString
-        )
     }
 }
